Quiz scenes
in these steps i added 10 more scenes to the game 4 for the body section
of the quiz, 3 for the head section of the quiz and 3 for the bones
section of the quiz (for each section there is a different scene
tv,mobile and pc). In these sceens there are animation of the ograns
(rotation,scale and postion animation ) i aslo added a timer of 30
secounds and also added a timer which adds 10 points which each right
answers, there are also 3 pick ups which are -1 wrong ans, bonus time
and the help button(which is still needs to be implemented).