Bonus scenes
At the end of the quiz a bonus game was created for the player to earn
some more valuable points at the end so he/she will more of a chance to
win the game. all the user have to do is to collect the organs in the
ice bucket and each organ they collect +1 will be added to the final
score.